requirements.txt
anthropic[vertex]
openai
google-genai
python-dotenv
fastapi
uvicorn[standard]
pyyaml
httpx
google-cloud-translate
